requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
Deprecated==1.2.13
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
packaging==21.3
py-cord==2.0.0b4
pyparsing==3.0.7
python-dotenv==0.19.2
redis==4.1.4
wrapt==1.13.3
xincraft==1.0.2
yarl==1.7.2
